+++ packages/vm/src/cache.rs
@@ -1,8 +1,9 @@
-use std::collections::HashSet;
+use std::collections::{BTreeSet, HashSet};
 use std::fs::{self, File, OpenOptions};
 use std::io::{Read, Write};
 use std::marker::PhantomData;
 use std::path::{Path, PathBuf};
+use std::str::FromStr;
 use std::sync::Mutex;
 use wasmer::{Engine, Store};
 
@@ -16,7 +17,7 @@ use crate::instance::{Instance, InstanceOptions};
 use crate::modules::{CachedModule, FileSystemCache, InMemoryCache, PinnedMemoryCache};
 use crate::parsed_wasm::ParsedWasm;
 use crate::size::Size;
-use crate::static_analysis::has_ibc_entry_points;
+use crate::static_analysis::{Entrypoint, ExportInfo, REQUIRED_IBC_EXPORTS};
 use crate::wasm_backend::{compile, make_compiling_engine, make_runtime_engine};
 
 const STATE_DIR: &str = "state";
@@ -117,8 +118,13 @@ pub struct Cache<A: BackendApi, S: Storage, Q: Querier> {
 
 #[derive(PartialEq, Eq, Debug)]
 pub struct AnalysisReport {
+    /// `true` if and only if all [`REQUIRED_IBC_EXPORTS`] exist as exported functions.
+    /// This does not guarantee they are functional or even have the correct signatures.
     pub has_ibc_entry_points: bool,
-    pub required_capabilities: HashSet<String>,
+    /// A set of all entrypoints that are exported by the contract.
+    pub entrypoints: BTreeSet<Entrypoint>,
+    /// The set of capabilities the contract requires.
+    pub required_capabilities: BTreeSet<String>,
 }
 
 impl<A, S, Q> Cache<A, S, Q>
@@ -274,9 +280,21 @@ where
         // Here we could use a streaming deserializer to slightly improve performance. However, this way it is DRYer.
         let wasm = self.load_wasm(checksum)?;
         let module = ParsedWasm::parse(&wasm)?;
+        let exports = module.exported_function_names(None);
+
+        let entrypoints = exports
+            .iter()
+            .filter_map(|export| Entrypoint::from_str(export).ok())
+            .collect();
+
         Ok(AnalysisReport {
-            has_ibc_entry_points: has_ibc_entry_points(&module),
-            required_capabilities: required_capabilities_from_module(&module),
+            has_ibc_entry_points: REQUIRED_IBC_EXPORTS
+                .iter()
+                .all(|required| exports.contains(required.as_ref())),
+            entrypoints,
+            required_capabilities: required_capabilities_from_module(&module)
+                .into_iter()
+                .collect(),
         })
     }
 
@@ -536,6 +554,7 @@ mod tests {
 
     static CONTRACT: &[u8] = include_bytes!("../testdata/hackatom.wasm");
     static IBC_CONTRACT: &[u8] = include_bytes!("../testdata/ibc_reflect.wasm");
+    static EMPTY_CONTRACT: &[u8] = include_bytes!("../testdata/empty.wasm");
     // Invalid because it doesn't contain required memory and exports
     static INVALID_CONTRACT_WAT: &str = r#"(module
         (type $t0 (func (param i32) (result i32)))
@@ -1274,6 +1293,7 @@ mod tests {
 
     #[test]
     fn analyze_works() {
+        use Entrypoint as E;
         let cache: Cache<MockApi, MockStorage, MockQuerier> =
             unsafe { Cache::new(make_stargate_testing_options()).unwrap() };
 
@@ -1283,20 +1303,42 @@ mod tests {
             report1,
             AnalysisReport {
                 has_ibc_entry_points: false,
-                required_capabilities: HashSet::new(),
+                entrypoints: BTreeSet::from([
+                    E::Instantiate,
+                    E::Migrate,
+                    E::Sudo,
+                    E::Execute,
+                    E::Query
+                ]),
+                required_capabilities: BTreeSet::new(),
             }
         );
 
         let checksum2 = cache.save_wasm(IBC_CONTRACT).unwrap();
         let report2 = cache.analyze(&checksum2).unwrap();
+        let mut ibc_contract_entrypoints =
+            BTreeSet::from([E::Instantiate, E::Migrate, E::Reply, E::Query]);
+        ibc_contract_entrypoints.extend(REQUIRED_IBC_EXPORTS);
         assert_eq!(
             report2,
             AnalysisReport {
                 has_ibc_entry_points: true,
-                required_capabilities: HashSet::from_iter([
+                entrypoints: ibc_contract_entrypoints,
+                required_capabilities: BTreeSet::from_iter([
                     "iterator".to_string(),
                     "stargate".to_string()
                 ]),
+            }
+        );
+
+        let checksum3 = cache.save_wasm(EMPTY_CONTRACT).unwrap();
+        let report3 = cache.analyze(&checksum3).unwrap();
+        assert_eq!(
+            report3,
+            AnalysisReport {
+                has_ibc_entry_points: false,
+                entrypoints: BTreeSet::new(),
+                required_capabilities: BTreeSet::from(["iterator".to_string()]),
             }
         );
     }